+++ rust/operator-binary/src/druid_controller.rs
@@ -17,6 +17,7 @@ use stackable_operator::{
         ConfigMapBuilder, ContainerBuilder, ObjectMetaBuilder, PodBuilder,
         PodSecurityContextBuilder, SecretOperatorVolumeSourceBuilder, VolumeBuilder,
     },
+    cluster_resources::ClusterResources,
     commons::{
         opa::OpaApiVersion,
         s3::{S3AccessStyle, S3ConnectionSpec},
@@ -33,7 +34,7 @@ use stackable_operator::{
     },
     kube::{
         runtime::{controller::Action, reflector::ObjectRef},
-        ResourceExt,
+        Resource, ResourceExt,
     },
     labels::{role_group_selector_labels, role_selector_labels},
     logging::controller::ReconcilerError,
@@ -50,7 +51,7 @@ use std::{
 };
 use strum::{EnumDiscriminants, IntoEnumIterator, IntoStaticStr};
 
-const FIELD_MANAGER_SCOPE: &str = "druidcluster";
+const CONTROLLER_NAME: &str = "druid-operator";
 
 pub struct Ctx {
     pub client: stackable_operator::client::Client,
@@ -61,6 +62,14 @@ pub struct Ctx {
 #[strum_discriminants(derive(IntoStaticStr))]
 #[allow(clippy::enum_variant_names)]
 pub enum Error {
+    #[snafu(display("failed to create cluster resources"))]
+    CreateClusterResources {
+        source: stackable_operator::error::Error,
+    },
+    #[snafu(display("failed to delete orphaned resources"))]
+    FinalizeClusterResources {
+        source: stackable_operator::error::Error,
+    },
     #[snafu(display("failed to apply global Service"))]
     ApplyRoleService {
         source: stackable_operator::error::Error,
@@ -235,10 +244,14 @@ pub async fn reconcile_druid(druid: Arc<DruidCluster>, ctx: Arc<Ctx>) -> Result<
     )
     .context(InvalidProductConfigSnafu)?;
 
+    let mut cluster_resources =
+        ClusterResources::new(APP_NAME, CONTROLLER_NAME, &druid.object_ref(&()))
+            .context(CreateClusterResourcesSnafu)?;
+
     for (role_name, role_config) in validated_role_config.iter() {
         let role_service = build_role_service(role_name, &druid)?;
-        client
-            .apply_patch(FIELD_MANAGER_SCOPE, &role_service, &role_service)
+        cluster_resources
+            .add(client, &role_service)
             .await
             .context(ApplyRoleServiceSnafu)?;
         for (rolegroup_name, rolegroup_config) in role_config.iter() {
@@ -264,20 +277,20 @@ pub async fn reconcile_druid(druid: Arc<DruidCluster>, ctx: Arc<Ctx>) -> Result<
                 rolegroup_config,
                 s3_conn.as_ref(),
             )?;
-            client
-                .apply_patch(FIELD_MANAGER_SCOPE, &rg_service, &rg_service)
+            cluster_resources
+                .add(client, &rg_service)
                 .await
                 .with_context(|_| ApplyRoleGroupServiceSnafu {
                     rolegroup: rolegroup.clone(),
                 })?;
-            client
-                .apply_patch(FIELD_MANAGER_SCOPE, &rg_configmap, &rg_configmap)
+            cluster_resources
+                .add(client, &rg_configmap)
                 .await
                 .with_context(|_| ApplyRoleGroupConfigSnafu {
                     rolegroup: rolegroup.clone(),
                 })?;
-            client
-                .apply_patch(FIELD_MANAGER_SCOPE, &rg_statefulset, &rg_statefulset)
+            cluster_resources
+                .add(client, &rg_statefulset)
                 .await
                 .with_context(|_| ApplyRoleGroupStatefulSetSnafu {
                     rolegroup: rolegroup.clone(),
@@ -286,15 +299,20 @@ pub async fn reconcile_druid(druid: Arc<DruidCluster>, ctx: Arc<Ctx>) -> Result<
     }
 
     // discovery
-    for discovery_cm in build_discovery_configmaps(&*druid, &*druid)
+    for discovery_cm in build_discovery_configmaps(&*druid, CONTROLLER_NAME, &*druid)
         .await
         .context(BuildDiscoveryConfigSnafu)?
     {
-        client
-            .apply_patch(FIELD_MANAGER_SCOPE, &discovery_cm, &discovery_cm)
+        cluster_resources
+            .add(client, &discovery_cm)
             .await
             .context(ApplyDiscoveryConfigSnafu)?;
     }
+
+    cluster_resources
+        .finalize(client)
+        .await
+        .context(FinalizeClusterResourcesSnafu)?;
 
     Ok(Action::await_change())
 }
@@ -313,7 +331,14 @@ pub fn build_role_service(role_name: &str, druid: &DruidCluster) -> Result<Servi
             .name(&role_svc_name)
             .ownerreference_from_resource(druid, None, Some(true))
             .context(ObjectMissingMetadataForOwnerRefSnafu)?
-            .with_recommended_labels(druid, APP_NAME, druid_version(druid)?, role_name, "global")
+            .with_recommended_labels(
+                druid,
+                APP_NAME,
+                druid_version(druid)?,
+                CONTROLLER_NAME,
+                role_name,
+                "global",
+            )
             .build(),
         spec: Some(ServiceSpec {
             ports: Some(vec![ServicePort {
@@ -417,6 +442,7 @@ fn build_rolegroup_config_map(
                 druid,
                 APP_NAME,
                 druid_version(druid)?,
+                CONTROLLER_NAME,
                 &rolegroup.role,
                 &rolegroup.role_group,
             )
@@ -452,6 +478,7 @@ fn build_rolegroup_services(
                 druid,
                 APP_NAME,
                 druid_version(druid)?,
+                CONTROLLER_NAME,
                 &rolegroup.role,
                 &rolegroup.role_group,
             )
@@ -512,6 +539,7 @@ fn build_rolegroup_statefulset(
             druid,
             APP_NAME,
             druid_version,
+            CONTROLLER_NAME,
             &rolegroup_ref.role,
             &rolegroup_ref.role_group,
         )
@@ -623,6 +651,7 @@ fn build_rolegroup_statefulset(
                 druid,
                 APP_NAME,
                 druid_version,
+                CONTROLLER_NAME,
                 &rolegroup_ref.role,
                 &rolegroup_ref.role_group,
             )